Make ENTER key work with 41CX 'CAT 2'
The ENTER key is special handled in the 41CX key parser. Added some code
to detect if ENTER is pressed when doing catalogs, and if we are on a
41CX machine, go to the appropriate 41CX code to make the catalog behave
as it should.

There only minor difference now is that the LCD becomes right justified
while it is searching for the next things to show in the catalog.