Rewrite the multiplication code
Make use of partial multiplication, 56x56, then two 2x56 and finally an
8x8 multiply.

This makes the code easier to read and understand, it should make it
faster but this has was not measured. It made the code a little bit
smaller compared to the previous attempts.

Overall it is slower than the HP-16C for a couple of reasons. Due to
register shortage we need to keep X register intact, previously it was
cleared as we took each nibble, making it possible to get out faster.
The DMUL final alignment of upper half result could be a lot faster if
we used nibble shifts (and bit shifts in the end if needed), but it
would cost more code.

It is possible to improve on this, but it requires code space and
finding a scratch register somehow, which is currently not desired, so
we live with the slower performance for now.