Rework value masking
Masking values for arithmetic operations is really needed. The previous
code was really focused on masking X, to some degree Y. Changed it so
that any value can be masked.

Starting on a way to load values (and mask) from a postfix operand byte,
stack registers are treated as 64-bit values, other status registers are
56 bits wide and then data memory is nibble sized based on current word
size. Unfortunately, this cause the code grow too big for a single 4K
page, so it is partly disabled for now.

Made it possible to obtain signs for X and Y and optionally make both X
and Y positive. This is needed for signed DMUL in order to get the
correct result in the upper half.

S9 is now more freely available to the code, previously it was a special
flag for word size 56, but this is seldom needed, and the few places
that need it (shifts) can deduce it from the M carry mask which will
have the special value 0 in this case.

Also ensure that the keyHandler end up in the first 1K of the module, as
this is required by the code that plants it for the key sequence parser.