Fix stack overrun bug
Bug was introduced when the combined FindBufferUserFlags_LiftStackS11
was added, unfortunately due to the page relative call mechanism, it
temporarily uses one stack level too much. Decided to fix it by aligning
chkbuf (which is the leaf routine called) to use the GSB256 mechanism
that does not do this, and is faster as well.