Change WINDOW key to be the DOT key
It feels a bit more natural in that extra characters outside the current
window are indicated by a dot, then use the dot key to show them.
It also has the benefit of being close to the 0-3 keys that are the most
likely keys to be used together with it.