Move MASKL and MASKR keys
Git the splendid(?) idea of putting MASKL and MASKR on shifted USER/PRGM
keys. They are sort of not used otherwise and since they are prompting
instructions it is kind of safe, just press back arrow to cancel the
operation. It is also kind of logical, one key to the left and one to
the right.