Fix accidental use of default keyboard handler
It turns out that the I/O vector is not entirely reliable as it is not
always called.

We rely on always having it called after an instruction complete and we
return to light sleep. However, if a new key is pressed fast enough, the
I/O poll vector is skipped and mainframe goes directly to the key
parser. When this happens, the keyboard takeover is not called and the
calculator reverts back to the default floating point behavior for that
key press.

This is of course very bad as anything can occasionally happen more or
less randomly.

To work around the problem, the service ROM vector is used. This adds
another page to the module image.