Make possible to use page 7
Page 7 is normally dedicated to the HPIL module.

Extended Functions (and HP-41CX) test its presence by a call to a
"random" address in it. Insert a NOP there to make this module appear as
not taking page 7 if it is put there. As long as everyone does the check
this way HP did, it is safe to use page 7.