Implement proper value IND arguments
Instructions that work on value numbers (not registers), can still be
indirect. Implemented the proper indirect behavior and fixed the code so
that we take care of any errors in the value before rearranging the
stack in any way. This to ensure correct behavior, an instruction that
results in an error have no other visible effect on the stack.