Make ROM header the first entry
Programs that do not have the module file header at hand will look at
the first entry in the FAT and interpret that as a ROM header. It is a
good idea to keep it that way, so the header and program literal XROMs
are now ordered in that way.