Fix lockup in program mode
The OS extension postpone the key down check to allow us to set the
environment properly. However, as I/O poll routines may reconstruct the
return address to mainframe (assuming it was called from there), it is
not for sure that we will return after the OS extension poll vector
call. If we do not, key presses will not be processed and the calculator
appears to be frozen.
To work around this, we check if a key is down before returning to the
I/O poll vector scanner. If a key is down, we stop doing I/O poll
vectors and go and handle the key instead. Which is what would have
happened normally if a key was down.