Use system digit entry flag
This is to make the behavior more similar to floating point digit entry
mode.

One particular case where this is needed is alarms which reset the digit
entry flag. While it actually seems to work for normal message alarms,
control alarms run programs which are likely to alter the stack. To
prevent editing of the number that the control alarm left in X, it is
best to terminate digit entry as the Time module suggests.