Correction to use of system digit entry
Key presses that does not result in changing the number being entered
and that does not terminate digit entry need to set up the keyboard take
over. The cases where this happen are shift/user and a digit entry that
is not accepted due to either too big number or outside selected base.