Improve digit entry performance
It is important that keyboard performance is good when entering or
deleting digits. Check if key has been released more often, as well as
earlier and later.

If the key has been released during processing and a new key is down at
the end, take a shortcut directly to handle the next key, just set up
the keyboard take over and go to WKUP20 to deal with the key.
Avoid updating the LCD, any buffer processing, just go for the new key
immediately.

The result is actually staggering. It is impossible for me to miss a
digit key if I have to key in different keys. Two successive identical
keys are never missed. Pressing the same key 4 times can be missed if I
manage to get them really fast, then it records 3. But I cannot always
manage that, it often gets all 4.

Comparing to floating point entry, it is clear that the integer mode is
better at recording digit entry keys!!! It is easier for me to miss 1 or
2 rapid key presses on the same key there.

The HP-16C using similar fast two-key entry does not perform better. It
is slightly behind on 4 identical key presses (more consistently miss a
key), but does equally well on other test. I can get it to miss on 4
different keys, but that is mostly thanks to the shorter throw on the
keys, the HP-41 have longer throw and does not type different keys as
fast.

In short, it is totally brutal! It is clearly better then the HP-41
floating point mode as well as very slightly ahead of the HP-16C in
integer mode.