Fix ADD/SUB overflow in unsigned mode
Made overflow flag behave as an overflow flag in unsigned mode. This is
consistent with the HP-16C. In reality it means that it behaves the same
as the carry flag. This is in contrast to most CPUs, where the overflow
flag works in signed mode. (CPUs do not tend to have signed and unsigned
modes, so it all kind of makes sense.)