Zero extend on increased word size
Changed the behavior when increasing word size. Always preserve
numerical values on stack by either sign or zero extending the
values. Previously only negative values were sign extended, but it
lead to confusion as register values with "garbage" in them would show
up. To avoid this, all numbers on the stack are always zero/sign
extended to preserve values.

Updated documentation on this matter.

Added a test case for implicit and explicit sign (and zero) extensions.