Remove use of Q register in WSIZE
There is no need to mask values in the stack, they will be masked
later as they are used. This means we do not need the new word mask
and can simplify the code by not building it and then get M register
free so we can skip using the Q register.