Improve performance on finding nibble registers
As the last valid data register address is 1FF, we do not need to
process beyond that point. Large numbers can actually take a while to
handle, so we try to get out quicker if the physical register address
overflows.

This actually saves some code size as well, with a minor speed penalty
on finding valid register addresses.