Rearrange code
Test for ALDI before testing if the window is set in G register. This
because ALDI does not set up the G register, making it read some
random data. In this case it does not matter as they go to the same
destination, but it is not entirely correct. Furthermore, if the code
is changed to make them differ, it may become a problem.