Correct display bug on large decimal numbers
Getting the digits when only displaying the first window did not
handle overflowing bits properly, and would in most cases end up not
thinking it was an overflow at all. This caused a large number being
entered (or displayed) with lots of zero digits in lower positions to
revert to show as just a single zero.

The result was visually very confusing and also wrong as significant
zero digits were not shown, nor were the overflow dot by the base
character shown.