Correct poll vector handling if bank 2 is left on
Bank 2 should not be used for poll vectors, but if they are we now
do the IO poll by enabling bank 1 and fall into the usual routine.

Leaving bank 2 on indicates a bug as it is not meant to happen. As far
as I can see Ladybug always revert back to bank 1 before giving
control back to mainframe. But it could at least happen if Ladybug is
bank switched together with another bank switch module that is not
well behaved.

For the turn-on poll vector it should not be needed as all banks
revert back to main bank as part of the calculator going to light or
deep sleep.