Always return with SS0 from ON poll vector
If the Ladybug buffer does not exist at power on, it would return to
the poll vector scan routine with something other than SS0 selected.
This violates the ROMCHK protocol and can lead to problems if another
power on poll vector routines assumed that SS0 was up (as it may do).