Add banked marker and fix module identity
The primary bank was not marked as banked, which would cause some
introspection module to think that it is not banked. I added the bank
switch routines in the standard place a while back so it makes sense
to mark it properly.

Fix the .moddesc file which had old version identifiers and fixed some
version issues in the code as well.