Use EXITAPP instead of FLOAT to leave Ladybug
Now that the shell mechanism allows for multiple applications it makes
sense to change the way we exit Ladybug. Previously there was a FLOAT
function to disable the Ladybug application, but it is now changed to
EXITAPP which simply drops the topmost application.

It is a good idea to have a consistent way of exiting an application
and then the only logical choice is to use a key that is not normally
used. The shifted USER key is a good choice as it normally has no
function bound to it and USER is about keyboard layouts.

Some minor rearrangements are needed for this as MASKL was on that key
previously. Move MASKL and MASKR right (PRGM and ALPHA), drop ALDI
would make more sense to have on a slightly change alpha keyboard that
currently is not defined.

This frees up the PI key position, so WSIZE? is put there, as it can
be useful in calculations and as a way to check the current
settings. Setting the word size is nearby, making it a sensible
location.