Correct backarrow during digit entry
Backarrow always executed as CLXI due to that internal digit entry
flag was not set up before testing for backarrow. It used to work in
the past, but that was probably due to that the IF_DigitEntry flag was
saved in the buffer, which seems to not happen now.

I have not investigated that further, but it does not need to be saved
in the buffer as we are obeying the system digit entry flag, the
IF_DigitEntry flag is just a local cached version of it. The use of
the shell is how we can separate digit entry in integer mode from the
ordinary digit entry.

It is desirable to obey the system flag as the Time module works with
that one when dealing with alarms.