Correct backarrow of single character
If the backarrow key is held down when having a single character
entered, CLXI is executed, but it can be NULLed by holding the key
down. Previously it would leave the single character number in X,
disable stack lift and terminate digit entry. Entering another number
then replaces it.

The built-in behavior is to enter 0 and enable stack lift, which I
believe is incorrect.

Now we put 0 in X and disable stack lift, which is more logical and
correct, but differs from the original behavior.

A more correct way would be to not leave digit entry, but that is
tricky to accomplish as the code that deals with NULL testing is out
of reach in mainframe and seems virtually impossible to implement, at
least if will not be easy. This is a fishy corner-case and I think
setting it to 0 and terminating digit entry is acceptable behavior.