Correct erasing all digits in program memory
Deleting all digits during digit entry in program mode now properly
removes the #LIT XROM instruction. This problem was most likely
introduced recently when doing the fix for rubbing out all characters
to leave 0 in X. As the value is now saved when erasing the last
digit, it takes care of not inserting a text literal to represent
0 (as it is not needed), so the erase everything digit entry in
program mode can be simplified to only deal with the left over #LIT
XROM instruction.