Fix and formalize stack lift enable
Took a look at stack lift enable. Try to follow the behavior of
similar built-in operations and apply a little bit of common sense.

It was mostly correct, but some functions did not enable stack lift
and left it untouched.

Add some automated tests for it.

Also document how functions affect the stack lift status.