Fix comment to avoid problems with preprocessor
The assembler uses the Clang preprocessor and it sometimes causes
issues with assembly comments as they are not comments in the eye of
the C preprocessor. This particular comment has a tick character which
the preprocessor thinks is a character literal and as it is not
terminated it will give a diagnostic message about it. Using a C style
comment prevents this problem.